Extract the old mapper
