+++ airbyte-server/src/test/java/io/airbyte/server/RequestLoggerTest.java
@@ -5,6 +5,7 @@
 package io.airbyte.server;
 
 import io.airbyte.commons.io.IOs;
+import io.airbyte.commons.server.RequestLogger;
 import io.airbyte.config.Configs.WorkerEnvironment;
 import io.airbyte.config.helpers.LogClientSingleton;
 import io.airbyte.config.helpers.LogConfigs;
